rename: `_template_utils` --> `_main_create`
